backend_fastapi/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
google-auth>=2.29.0
httpx>=0.27.0
requests>=2.31.0
python-multipart>=0.0.9
